Removed tiny-sdf submodule
