+++ .gitmodules
@@ -1,3 +1,1 @@
-[submodule "external/tiny-sdf"]
-	path = external/tiny-sdf
-	url = git@github.com:Zemledelec/tiny-sdf.git
+
