+++ examples/billboardsDragging/billboardsDragging.html
@@ -34,9 +34,9 @@
                 'lonlat': [8.25, 46.74],
                 'label': {
                     face: "Roboto-Regular",
-                    'text': 'Touch,me. !HelLo@ 1287',
-                    'size': 24,
-                    color: "rgba(0,0,0,255)",
+                    'text': 'Hello, world! Hi, Globus$12xXoOopgjl',
+                    'size': 11,
+                    color: "rgba(0,0,255,255)",
                     'outlineColor': "rgba(255,255,255,1)",
                     'outline': 0
                 }
