+++ sandbox/temp/temp.html
@@ -1,0 +1,10 @@
+<head>
+    <title>OpenGlobus - Earth planet</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <script type="module" src="./rays.js"></script>
+    <link rel="stylesheet" href="../../css/og.css" type="text/css" />
+</head>
+<body>
+    <div id="globus" style="width:100%;height:100%"></div>
+</body>
+</html>
